debezium/dbz#2376 Move the JSON string formatting mode into the MySQL connector
