+++ debezium-connector-binlog/src/main/java/io/debezium/connector/binlog/BinlogConnectorConfig.java
@@ -121,51 +121,6 @@ public abstract class BinlogConnectorConfig extends HistorizedRelationalDatabase
     }
 
     /**
-     * Set of predefined JsonStringFormattingMode options.
-     */
-    public enum JsonStringFormattingMode implements EnumeratedValue {
-        /**
-         * Serializes {@code JSON} values read from the binlog without whitespace, e.g. {@code {"key":"value"}}.
-         */
-        LEGACY("legacy"),
-        /**
-         * Serializes {@code JSON} values read from the binlog as the database server does when the column is
-         * read via JDBC, e.g. {@code {"key": "value"}}.
-         */
-        DATABASE("database");
-
-        private final String value;
-
-        JsonStringFormattingMode(String value) {
-            this.value = value;
-        }
-
-        @Override
-        public String getValue() {
-            return value;
-        }
-
-        /**
-         * Determine if the supplied value is one of the predefined options.
-         *
-         * @param value the configuration property value; may not be null
-         * @return the matching option, or null if no match is found
-         */
-        public static JsonStringFormattingMode parse(String value) {
-            if (value == null) {
-                return null;
-            }
-            value = value.trim();
-            for (JsonStringFormattingMode option : JsonStringFormattingMode.values()) {
-                if (option.getValue().equalsIgnoreCase(value)) {
-                    return option;
-                }
-            }
-            return null;
-        }
-    }
-
-    /**
      * Set of predefined SnapshotLockingMode options or aliases.
      */
     public enum SnapshotLockingMode implements EnumeratedValue {
@@ -571,17 +526,6 @@ public abstract class BinlogConnectorConfig extends HistorizedRelationalDatabase
                     + "'precise' uses java.math.BigDecimal to represent values, which are encoded in the change events using a binary representation and Kafka Connect's 'org.apache.kafka.connect.data.Decimal' type; "
                     + "'long' (the default) represents values using Java's 'long', which may not offer the precision but will be far easier to use in consumers.");
 
-    public static final Field JSON_STRING_FORMATTING_MODE = Field.create("json.string.formatting.mode")
-            .withDisplayName("JSON string formatting mode")
-            .withEnum(JsonStringFormattingMode.class, JsonStringFormattingMode.LEGACY)
-            .withWidth(ConfigDef.Width.SHORT)
-            .withImportance(ConfigDef.Importance.LOW)
-            .withGroup(Field.createGroupEntry(Field.Group.CONNECTOR))
-            .withDescription("Specify how JSON values read from the binlog are serialized in change events, including: "
-                    + "'legacy' (the default) uses a compact format without whitespace, e.g. {\"key\":\"value\"}; "
-                    + "'database' uses the format the database server produces when the column is read via JDBC, e.g. {\"key\": \"value\"}, "
-                    + "so that streamed values are consistent with snapshotted ones.");
-
     /**
      * @deprecated Use {@link #EVENT_PROCESSING_FAILURE_HANDLING_MODE} instead, will be removed in Debezium 3.0.
      */
@@ -714,8 +658,8 @@ public abstract class BinlogConnectorConfig extends HistorizedRelationalDatabase
             .group(Field.Group.CONNECTION_ADVANCED, ON_CONNECT_STATEMENTS, SERVER_ID_OFFSET, CONNECTION_TIMEOUT_MS, KEEP_ALIVE, KEEP_ALIVE_INTERVAL_MS,
                     KEEP_ALIVE_MAX_RECONNECT_ATTEMPTS, USE_NONGRACEFUL_DISCONNECT, BINLOG_NET_WRITE_TIMEOUT, BINLOG_NET_READ_TIMEOUT)
             .group(Field.Group.CONNECTION_ADVANCED_SSL, SSL_KEYSTORE, SSL_KEYSTORE_PASSWORD, SSL_TRUSTSTORE, SSL_TRUSTSTORE_PASSWORD)
-            .group(Field.Group.CONNECTOR, BIGINT_UNSIGNED_HANDLING_MODE, JSON_STRING_FORMATTING_MODE, TIME_PRECISION_MODE, ENABLE_TIME_ADJUSTER,
-                    SCHEMA_NAME_ADJUSTMENT_MODE, GTID_SOURCE_INCLUDES, GTID_SOURCE_EXCLUDES, GTID_SOURCE_FILTER_DML_EVENTS)
+            .group(Field.Group.CONNECTOR, BIGINT_UNSIGNED_HANDLING_MODE, TIME_PRECISION_MODE, ENABLE_TIME_ADJUSTER, SCHEMA_NAME_ADJUSTMENT_MODE, GTID_SOURCE_INCLUDES,
+                    GTID_SOURCE_EXCLUDES, GTID_SOURCE_FILTER_DML_EVENTS)
             .group(Field.Group.CONNECTOR_ADVANCED, ROW_COUNT_FOR_STREAMING_RESULT_SETS, BUFFER_SIZE_FOR_BINLOG_READER, INCLUDE_SQL_QUERY, IGNORE_GTID_ON_RECOVERY)
             .group(Field.Group.CONNECTOR_SNAPSHOT, SNAPSHOT_MODE, SNAPSHOT_QUERY_MODE, SNAPSHOT_QUERY_MODE_CUSTOM_NAME, INCREMENTAL_SNAPSHOT_CHUNK_SIZE,
                     INCREMENTAL_SNAPSHOT_ALLOW_SCHEMA_CHANGES)
@@ -729,7 +673,6 @@ public abstract class BinlogConnectorConfig extends HistorizedRelationalDatabase
     private final Duration connectionTimeout;
     private final EventProcessingFailureHandlingMode inconsistentSchemaFailureHandlingMode;
     private final BigIntUnsignedHandlingMode bigIntUnsignedHandlingMode;
-    private final JsonStringFormattingMode jsonStringFormattingMode;
     private final boolean readOnlyConnection;
     private final boolean ignoreGtidOnRecovery;
     private final boolean resolveLikeTableSchema;
@@ -758,7 +701,6 @@ public abstract class BinlogConnectorConfig extends HistorizedRelationalDatabase
         this.connectionTimeout = Duration.ofMillis(config.getLong(CONNECTION_TIMEOUT_MS));
         this.inconsistentSchemaFailureHandlingMode = EventProcessingFailureHandlingMode.parse(config.getString(INCONSISTENT_SCHEMA_HANDLING_MODE));
         this.bigIntUnsignedHandlingMode = BigIntUnsignedHandlingMode.parse(config.getString(BIGINT_UNSIGNED_HANDLING_MODE));
-        this.jsonStringFormattingMode = JsonStringFormattingMode.parse(config.getString(JSON_STRING_FORMATTING_MODE));
         this.ignoreGtidOnRecovery = config.getBoolean(IGNORE_GTID_ON_RECOVERY);
         this.resolveLikeTableSchema = config.getBoolean(RESOLVE_LIKE_TABLE_SCHEMA);
     }
@@ -811,13 +753,6 @@ public abstract class BinlogConnectorConfig extends HistorizedRelationalDatabase
      */
     public BigIntUnsignedHandlingMode getBigIntUnsignedHandlingMode() {
         return bigIntUnsignedHandlingMode;
-    }
-
-    /**
-     * @return the {@code JSON} string formatting mode
-     */
-    public JsonStringFormattingMode getJsonStringFormattingMode() {
-        return jsonStringFormattingMode;
     }
 
     /**
