+++ debezium-connector-binlog/src/main/java/io/debezium/connector/binlog/jdbc/BinlogValueConverters.java
@@ -45,7 +45,6 @@ import io.debezium.DebeziumException;
 import io.debezium.annotation.Immutable;
 import io.debezium.config.CommonConnectorConfig.BinaryHandlingMode;
 import io.debezium.config.CommonConnectorConfig.EventConvertingFailureHandlingMode;
-import io.debezium.connector.binlog.BinlogConnectorConfig.JsonStringFormattingMode;
 import io.debezium.connector.binlog.BinlogGeometry;
 import io.debezium.connector.binlog.BinlogUnsignedIntegerConverter;
 import io.debezium.connector.binlog.charset.BinlogCharsetRegistry;
@@ -118,7 +117,6 @@ public abstract class BinlogValueConverters extends JdbcValueConverters {
 
     private final EventConvertingFailureHandlingMode eventConvertingFailureHandlingMode;
     private final BinlogCharsetRegistry charsetRegistry;
-    private final JsonStringFormattingMode jsonStringFormattingMode;
     private final byte[] unavailableValuePlaceholderBinary;
     private final String unavailableValuePlaceholderString;
 
@@ -132,8 +130,6 @@ public abstract class BinlogValueConverters extends JdbcValueConverters {
      * @param binaryHandlingMode how binary columns should be treated
      * @param adjuster a temporal adjuster to make a database specific time before conversion
      * @param eventConvertingFailureHandlingMode how to handle conversion failures
-     * @param jsonStringFormattingMode how {@code JSON} values read from the binlog are serialized; may be null,
-     *            which is treated as {@link JsonStringFormattingMode#LEGACY}
      * @param serviceRegistry the service registry instance, should not be {@code null}
      * @param unavailableValuePlaceholder the placeholder bytes for unavailable column values; may be null
      */
@@ -143,12 +139,10 @@ public abstract class BinlogValueConverters extends JdbcValueConverters {
                                  BinaryHandlingMode binaryHandlingMode,
                                  TemporalAdjuster adjuster,
                                  EventConvertingFailureHandlingMode eventConvertingFailureHandlingMode,
-                                 JsonStringFormattingMode jsonStringFormattingMode,
                                  ServiceRegistry serviceRegistry,
                                  byte[] unavailableValuePlaceholder) {
         super(decimalMode, temporalPrecisionMode, ZoneOffset.UTC, adjuster, bigIntUnsignedMode, binaryHandlingMode);
         this.eventConvertingFailureHandlingMode = eventConvertingFailureHandlingMode;
-        this.jsonStringFormattingMode = jsonStringFormattingMode;
         this.charsetRegistry = serviceRegistry.getService(BinlogCharsetRegistry.class);
         this.unavailableValuePlaceholderBinary = unavailableValuePlaceholder;
         this.unavailableValuePlaceholderString = unavailableValuePlaceholder != null ? new String(unavailableValuePlaceholder) : null;
@@ -166,7 +160,7 @@ public abstract class BinlogValueConverters extends JdbcValueConverters {
                                  EventConvertingFailureHandlingMode eventConvertingFailureHandlingMode,
                                  ServiceRegistry serviceRegistry) {
         this(decimalMode, temporalPrecisionMode, bigIntUnsignedMode, binaryHandlingMode, adjuster,
-                eventConvertingFailureHandlingMode, null, serviceRegistry, null);
+                eventConvertingFailureHandlingMode, serviceRegistry, null);
     }
 
     @Override
@@ -491,13 +485,13 @@ public abstract class BinlogValueConverters extends JdbcValueConverters {
         return convertValue(column, fieldDefn, data, "{}", (r) -> {
             if (data instanceof byte[]) {
                 // The BinlogReader sees these JSON values as binary encoded, so we use the binlog client library's utility
-                // to parse the database's internal binary representation into a JSON string, using the configured formatter.
+                // to parse the database's internal binary representation into a JSON string, using the standard formatter.
                 if (((byte[]) data).length == 0) {
                     r.deliver(column.isOptional() ? null : "{}");
                 }
                 else {
                     try {
-                        r.deliver(parseJsonBinary((byte[]) data));
+                        r.deliver(JsonBinary.parseAsString((byte[]) data));
                     }
                     catch (IOException e) {
                         if (eventConvertingFailureHandlingMode == FAIL) {
@@ -516,25 +510,6 @@ public abstract class BinlogValueConverters extends JdbcValueConverters {
                 r.deliver(data);
             }
         });
-    }
-
-    private String parseJsonBinary(byte[] data) throws IOException {
-        if (jsonStringFormattingMode == JsonStringFormattingMode.DATABASE && !isTextJson(data)) {
-            DatabaseJsonStringFormatter formatter = new DatabaseJsonStringFormatter();
-            JsonBinary.parse(data, formatter);
-            return formatter.getString();
-        }
-        return JsonBinary.parseAsString(data);
-    }
-
-    /**
-     * MariaDB stores {@code JSON} columns as {@code LONGTEXT} and replicates them as text, which
-     * {@link JsonBinary#parseAsString(byte[])} passes through unchanged; the same check is applied here so
-     * that text values do not reach the binary parser.
-     */
-    private static boolean isTextJson(byte[] data) {
-        // Binary JSON type markers are all <= 0x0f
-        return data[0] > 0x0f;
     }
 
     /**
