+++ debezium-connector-binlog/src/test/java/io/debezium/connector/binlog/BinlogValueConvertersTest.java
@@ -11,7 +11,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.math.BigDecimal;
-import java.nio.charset.StandardCharsets;
 import java.sql.Timestamp;
 import java.time.Duration;
 import java.time.LocalDate;
@@ -28,7 +27,6 @@ import org.junit.jupiter.api.Test;
 import io.debezium.DebeziumException;
 import io.debezium.config.CommonConnectorConfig.BinaryHandlingMode;
 import io.debezium.config.CommonConnectorConfig.EventConvertingFailureHandlingMode;
-import io.debezium.config.Configuration;
 import io.debezium.connector.binlog.event.BinlogDateTimeValue;
 import io.debezium.connector.binlog.event.BinlogDateValue;
 import io.debezium.connector.binlog.jdbc.BinlogValueConverters;
@@ -56,75 +54,6 @@ public abstract class BinlogValueConvertersTest<C extends SourceConnector> imple
     private static final TemporalAdjuster ADJUSTER = BinlogValueConverters::adjustTemporal;
     private static final byte[] INVALID_JSON = { 2, 1, 0, 91, 0, 0, 7, 0, 2, 0, 84, 0, 18, 0, 4, 0, 22, 0, 6, 0, 12, 28,
             0, 0, 47, 0, 116, 121, 112, 101 };
-
-    // Binary (binlog) representations of JSON values, in the MySQL binary JSON format
-
-    // {"key1": "value1", "key2": "value2"}
-    private static final byte[] BINARY_JSON_OBJECT = binaryJson(
-            0x00, // small object
-            0x02, 0x00, // element count
-            0x28, 0x00, // size in bytes
-            0x12, 0x00, 0x04, 0x00, // key "key1": offset, length
-            0x16, 0x00, 0x04, 0x00, // key "key2": offset, length
-            0x0c, 0x1a, 0x00, // value "value1": string, offset
-            0x0c, 0x21, 0x00, // value "value2": string, offset
-            'k', 'e', 'y', '1',
-            'k', 'e', 'y', '2',
-            0x06, 'v', 'a', 'l', 'u', 'e', '1',
-            0x06, 'v', 'a', 'l', 'u', 'e', '2');
-
-    // [1, 2]
-    private static final byte[] BINARY_JSON_ARRAY = binaryJson(
-            0x02, // small array
-            0x02, 0x00, // element count
-            0x0a, 0x00, // size in bytes
-            0x05, 0x01, 0x00, // inlined int16 1
-            0x05, 0x02, 0x00); // inlined int16 2
-
-    // {"a": [1, 2], "b": {"c": "d"}}
-    private static final byte[] BINARY_JSON_NESTED = binaryJson(
-            0x00, // small object
-            0x02, 0x00, // element count
-            0x2c, 0x00, // size in bytes
-            0x12, 0x00, 0x01, 0x00, // key "a": offset, length
-            0x13, 0x00, 0x01, 0x00, // key "b": offset, length
-            0x02, 0x14, 0x00, // value [1, 2]: small array, offset
-            0x00, 0x1e, 0x00, // value {"c": "d"}: small object, offset
-            'a',
-            'b',
-            0x02, 0x00, 0x0a, 0x00, // nested array header
-            0x05, 0x01, 0x00,
-            0x05, 0x02, 0x00,
-            0x01, 0x00, 0x0e, 0x00, // nested object header
-            0x0b, 0x00, 0x01, 0x00, // key "c": offset, length
-            0x0c, 0x0c, 0x00, // value "d": string, offset
-            'c',
-            0x01, 'd');
-
-    // {}
-    private static final byte[] BINARY_JSON_EMPTY_OBJECT = binaryJson(0x00, 0x00, 0x00, 0x04, 0x00);
-
-    // []
-    private static final byte[] BINARY_JSON_EMPTY_ARRAY = binaryJson(0x02, 0x00, 0x00, 0x04, 0x00);
-
-    // "tab\tand\nnewline"
-    private static final byte[] BINARY_JSON_ESCAPED_SCALAR = binaryJson(
-            0x0c, 0x0f, // string scalar, length
-            't', 'a', 'b', 0x09, 'a', 'n', 'd', 0x0a, 'n', 'e', 'w', 'l', 'i', 'n', 'e');
-
-    // 2015-01-15 23:24:25 as a DATETIME scalar
-    private static final byte[] BINARY_JSON_DATETIME_SCALAR = binaryJsonPackedTemporal(
-            0x0c, packDatetime(2015, 1, 15, 23, 24, 25), 0);
-
-    // 23:24:25.12 as a TIME scalar
-    private static final byte[] BINARY_JSON_TIME_SCALAR = binaryJsonPackedTemporal(
-            0x0b, (23L << 12) | (24L << 6) | 25L, 120_000);
-
-    // x'cafe' as an opaque scalar
-    private static final byte[] BINARY_JSON_BLOB_SCALAR = binaryJson(
-            0x0f, 0x0f, // opaque scalar, MYSQL_TYPE_VARCHAR
-            0x02, // length
-            0xca, 0xfe);
 
     @Test
     void shouldAdjustLocalDateWithTwoDigitYears() {
@@ -196,97 +125,6 @@ public abstract class BinlogValueConvertersTest<C extends SourceConnector> imple
         assertThat(converters.converter(colB, fieldB).convert(null)).isEqualTo("{}");
         assertThat(converters.converter(colB, fieldB).convert("{ \"key1\": \"val1\", \"key2\": {\"key3\":\"val3\"} }"))
                 .isEqualTo("{ \"key1\": \"val1\", \"key2\": {\"key3\":\"val3\"} }");
-    }
-
-    @Test
-    @FixFor("debezium/dbz#2376")
-    public void testJsonValuesStreamingDiffersFromSnapshotWithLegacyFormattingMode() {
-        ValueConverter converter = jsonColumnConverter(BinlogConnectorConfig.JsonStringFormattingMode.LEGACY);
-
-        // Snapshot values are the server text, delivered unchanged
-        assertThat(converter.convert("{\"key1\": \"value1\", \"key2\": \"value2\"}")).isEqualTo("{\"key1\": \"value1\", \"key2\": \"value2\"}");
-        assertThat(converter.convert("[1, 2]")).isEqualTo("[1, 2]");
-
-        // The same documents read from the binlog are serialized differently
-        assertThat(converter.convert(BINARY_JSON_OBJECT)).isEqualTo("{\"key1\":\"value1\",\"key2\":\"value2\"}");
-        assertThat(converter.convert(BINARY_JSON_ARRAY)).isEqualTo("[1,2]");
-        assertThat(converter.convert(BINARY_JSON_NESTED)).isEqualTo("{\"a\":[1,2],\"b\":{\"c\":\"d\"}}");
-        assertThat(converter.convert(BINARY_JSON_DATETIME_SCALAR)).isEqualTo("\"2015-01-15 23:24:25\"");
-        assertThat(converter.convert(BINARY_JSON_TIME_SCALAR)).isEqualTo("\"23:24:25.12\"");
-        assertThat(converter.convert(BINARY_JSON_BLOB_SCALAR)).isEqualTo("\"yv4=\"");
-    }
-
-    @Test
-    @FixFor("debezium/dbz#2376")
-    public void testJsonValuesStreamingMatchesSnapshotWithDatabaseFormattingMode() {
-        ValueConverter converter = jsonColumnConverter(BinlogConnectorConfig.JsonStringFormattingMode.DATABASE);
-
-        assertThat(converter.convert("{\"key1\": \"value1\", \"key2\": \"value2\"}")).isEqualTo("{\"key1\": \"value1\", \"key2\": \"value2\"}");
-        assertThat(converter.convert("[1, 2]")).isEqualTo("[1, 2]");
-
-        // Values read from the binlog now serialize to the strings the server produces
-        assertThat(converter.convert(BINARY_JSON_OBJECT)).isEqualTo("{\"key1\": \"value1\", \"key2\": \"value2\"}");
-        assertThat(converter.convert(BINARY_JSON_ARRAY)).isEqualTo("[1, 2]");
-        assertThat(converter.convert(BINARY_JSON_NESTED)).isEqualTo("{\"a\": [1, 2], \"b\": {\"c\": \"d\"}}");
-        assertThat(converter.convert(BINARY_JSON_DATETIME_SCALAR)).isEqualTo("\"2015-01-15 23:24:25.000000\"");
-        assertThat(converter.convert(BINARY_JSON_TIME_SCALAR)).isEqualTo("\"23:24:25.120000\"");
-        assertThat(converter.convert(BINARY_JSON_BLOB_SCALAR)).isEqualTo("\"base64:type15:yv4=\"");
-
-        // Documents without entry separators are unaffected, and escaping is inherited unchanged
-        assertThat(converter.convert(BINARY_JSON_EMPTY_OBJECT)).isEqualTo("{}");
-        assertThat(converter.convert(BINARY_JSON_EMPTY_ARRAY)).isEqualTo("[]");
-        assertThat(converter.convert(BINARY_JSON_ESCAPED_SCALAR)).isEqualTo("\"tab\\tand\\nnewline\"");
-
-        // MariaDB replicates JSON as text, which is still passed through unchanged
-        assertThat(converter.convert("{\"a\":2,\"b\": 3}".getBytes(StandardCharsets.UTF_8))).isEqualTo("{\"a\":2,\"b\": 3}");
-    }
-
-    private ValueConverter jsonColumnConverter(BinlogConnectorConfig.JsonStringFormattingMode jsonStringFormattingMode) {
-        Configuration configuration = Configuration.create()
-                .with(BinlogConnectorConfig.JSON_STRING_FORMATTING_MODE, jsonStringFormattingMode)
-                .build();
-        BinlogValueConverters converters = getValueConverters(configuration);
-
-        DdlParser parser = getDdlParser();
-        Tables tables = new Tables();
-        parser.parse("CREATE TABLE JSON_TABLE (A JSON);", tables);
-        Table table = tables.forTable(new TableId(null, null, "JSON_TABLE"));
-
-        Column column = table.columnWithName("A");
-        Field field = new Field(column.name(), -1, converters.schemaBuilder(column).optional().build());
-        return converters.converter(column, field);
-    }
-
-    private static byte[] binaryJson(int... bytes) {
-        byte[] result = new byte[bytes.length];
-        for (int i = 0; i < bytes.length; i++) {
-            result[i] = (byte) bytes[i];
-        }
-        return result;
-    }
-
-    /**
-     * Packs a datetime as MySQL does: 17 bits of {@code year * 13 + month}, then 5 bits each for the day
-     * and hour and 6 bits each for the minute and second.
-     */
-    private static long packDatetime(int year, int month, int day, int hour, int min, int sec) {
-        return (((long) year * 13 + month) << 22) | ((long) day << 17) | ((long) hour << 12) | ((long) min << 6) | sec;
-    }
-
-    /**
-     * Builds a JSON temporal scalar: an opaque value holding the 8-byte little-endian packed temporal
-     * value, whose lowest 24 bits are the microseconds.
-     */
-    private static byte[] binaryJsonPackedTemporal(int columnType, long packedValue, int microSeconds) {
-        long raw = (packedValue << 24) | microSeconds;
-        byte[] result = new byte[11];
-        result[0] = 0x0f; // opaque scalar
-        result[1] = (byte) columnType;
-        result[2] = 0x08; // length
-        for (int i = 0; i < 8; i++) {
-            result[3 + i] = (byte) (raw >> (8 * i));
-        }
-        return result;
     }
 
     @Test
@@ -676,8 +514,6 @@ public abstract class BinlogValueConvertersTest<C extends SourceConnector> imple
                                                                 BinaryHandlingMode binaryHandlingMode,
                                                                 TemporalAdjuster temporalAdjuster,
                                                                 EventConvertingFailureHandlingMode eventConvertingFailureHandlingMode);
-
-    protected abstract BinlogValueConverters getValueConverters(Configuration configuration);
 
     protected abstract DdlParser getDdlParser();
