+++ debezium-connector-mariadb/src/main/java/io/debezium/connector/mariadb/MariaDbConnectorTask.java
@@ -338,7 +338,6 @@ public class MariaDbConnectorTask extends BinlogSourceTask<MariaDbPartition, Mar
                 connectorConfig.binaryHandlingMode(),
                 connectorConfig.isTimeAdjustedEnabled() ? MariaDbValueConverters::adjustTemporal : x -> x,
                 connectorConfig.getEventConvertingFailureHandlingMode(),
-                connectorConfig.getJsonStringFormattingMode(),
                 connectorConfig.getServiceRegistry());
     }
