+++ debezium-connector-mysql/src/main/java/io/debezium/connector/mysql/jdbc/DatabaseJsonStringFormatter.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.connector.binlog.jdbc;
+package io.debezium.connector.mysql.jdbc;
 
 import java.util.Base64;
 import java.util.Locale;
